requirements.txt
pip~=21.0.1
Django~=3.1
python-decouple~=3.3
djangorestframework~= 3.12
mysqlclient~=2.0
django-filter~=2.4
djangorestframework-simplejwt~=4.6.0
wheel~=0.36.1
lxml~=4.3.5
setuptools~=51.0.0
future~=0.18.2
celery~=5.0.5
youtube_dl~=2021.4.17
django-celery-results~=2.0.1
coreapi~=2.3.3
drf-yasg~=1.20.0
kombu~=5.0.2
